pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Heric", email = "hericmdms@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "drf-spectacular>=0.26.4",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm]

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

